Feat: Implemented upper cutoffs (radial/vdw)
Implemented upper cutoffs (outer) radial and upper scaled vdw cutoffs, in the
form of two new arguments --upper-rc and --upper-vdw, which are
the complements of the --rc and --vdw cutoff arguments. That is, they
exclude points outside a given cutoff radius from the host framework
atoms from being sampled. As the previously existing arguments, the cutoffs can be
either in Ångström, or in terms of a scaling factor of the vdW radius.