Fix: Use unit cell for making grid when present, and fixed path handling , in analyze_batch_reftraj.py.
Fixed so that the structure_unitcell.cif file is read when present, i.e.
when a sampling supercell has been used with --ssc or --mic-cutoff, when
compiling the grid with analyze_batch_reftraj.py. Also improved the path
handling in this script, so that it is no longer necessary to include
trailing slashes of batch directory path-arguments on the commandline.